Update the things

See merge request Team6479/Robot2020!13

+++ src/main/java/frc/robot/subsystems/Turret.java
@@ -178,7 +178,8 @@ public class Turret extends SubsystemBase {
   }
 
   public int getPIDError() {
-    return motor.getClosedLoopError();
+    // TODO: why (int)
+    return (int)motor.getClosedLoopError();
   }
